requirements/base.txt
Django==5.2.7
django-environ==0.12.0
psycopg2==2.9.11
djangorestframework==3.16.1

celery==5.5.3
django-celery-results==2.6.0
django-celery-beat==2.8.1

whitenoise==6.11.0

pgvector==0.4.2

django-filter==25.2
django-extensions==4.1
django-cors-headers==4.9.0
django-storages==1.14.6
django-structlog[celery]==9.1.1
django-redis==6.0.0
drf-spectacular==0.29.0

drf-jwt==1.19.2

boto3==1.40.52
attrs==25.4.0

gunicorn==23.0.0
sentry-sdk==2.42.0

requests==2.32.5
httpx==0.28.1
PyYAML==6.0.3
Pillow>=10.0.0
tenacity==9.1.4
respx==0.22.0
beautifulsoup4==4.14.3
#google-api-python-client==2.184.0
#google-auth==2.41.1
#google-auth-httplib2==0.2.0
#google-auth-oauthlib==1.2.2

pyotp==2.9.0
qrcode==8.2
tabulate==0.9.0
